RF: list all extractors/indexers with distribution
Since distributions may overwrite extractors or indexers
with new implementations, we add the distribution name
to the reported names. That ensures (with the current
implementation) that all extractors and indexers are
reported. In addition we report the distribution for
the different extractors and indexers.

It should be noted, that extractors and indexers from
different distributions are not necessarily compatible,
e.g. extractors from datalad-metalad are not compatible
with datalad-core extractor invocations. Therefore we
use a built-in fall back indexer if multiple indexers
are detected. The new wtf-format should simplify
debugging of this situation.