OPT: resolve_path() can take pre-resolved dataset instance
This disentangled the ability to switch behavior depending on dataset arg
type, from the necessity to repeatedly run `require_dataset()` when
resolve_path()` is called repeatedly. It is common to call this function
in a loop on every single input argument. So with lots of arguments this
will have a performance impact.