OPT: Minimize checks when root path for sudatasets() call is dataset root
https://github.com/datalad/datalad/pull/4868#issuecomment-688417401
reported substantial time difference with and without `--dataset` on
datasets with many subdatasets. The cause was that without `--dataset`
`os.curdir` was used as a query path, whereas with `--dataset`, no explicit
query path was used. Each query path adds runtime from various safety
checks. However, when we generate the query path, we can optimze for
that.

This change uses no query path, if the to be queried dataset has its
current in `PWD`. This will yield the same results, but bypass all
path-related processing. As a consequence, command performance is the
same with or without as `dataset` argument given.