RF: dissolve _add_meta - do in _add_urls
Initial idea was to just do it while working on a row. Unfortunately if we do that
annex would start "ignoring" always_commit=False, so I needed to do a separate loop.

But I think it would still be benefitial to have it absorbed within the same function
so we could just collect necessary filename: meta  records and not to loop through the
list of rows again to discover if we need to add any meta or not.

Also so far there is an anecdotal evidence that it seems to resolve jumping
total progress bar, seems to advance nicely upon each file