ENH: graceful interruption of parallel execution
I was always wondering, what is this "offlineimap" wants from me when
I press Ctrl-C and it still keeps going asking for more Ctrl-C or just
wait a little.  At times I was patient and waited...

In our case I think it might be crucial to avoid data loss in other
parallel processes if one process dies - we should not jump to
interrupt all the rest, but we also unlikely want to proceed
forward (although we might decide to add that feature, so to become
"friendlier" to on_failure="ignore" alike invocation.

But also, if user interrupts with Ctrl-C, and he/she can wait for
already running processes to complete, it would be great!  So I
somewhat mimiced two-stage killing procedure of offlineimap.  Upon
initial exception (might come from producer or consumer or user,
i.e. Ctrl-C) we just purge the "queues" so no new jobs are to be ran,
but we let already running ones to finish.  If there is a subsequent
Ctrl-C (all other exceptions are just logged), then we do a
forceful (well, as much as we can) shutdown.

I played with that script of mine on top of addurls
    #!/bin/bash

    export PS4='> '

    set -eu
    cd "$(mktemp -d ${TMPDIR:-/tmp}/dl-XXXXXXX)"
    topd=`pwd`

    echo -e "url,filename,bogus1,bogus2" > list.csv
    mkdir src
    n=${1:-2}
    J=${2:-10}
    (cd src;
    for id in `seq 1 $n`; do
        d=d$id
        mkdir $d
        for s in `seq 1 $n`; do
            f="$d//$s"
            fu="$d/$s"
            # so there is at least something notable for annex to checksum
            #dd if=/dev/random of=$fu count=10000 >& /dev/null
            # super fast one ;)
            echo $s > $fu;
            echo "file:///$topd/src/$fu,$f,pocus,focus" >> ../list.csv;
        done
    done
    )

    set -x
    datalad -c annex.security.allowed-url-schemes=file create dest
    (
    cd dest;
    time datalad -c annex.security.allowed-url-schemes=file --on-failure stop addurls -J$J ../list.csv '{url}' '{filename}'

    )

and I liked how it behaved -- you could see some additional results
yielded/reported while no new runners are executed etc.

Additional minor enhancement - producer could be a generator function
now (one use case test added)