TST: appveyor: Restore execution of modules that use with_testrepos
A good number of test modules no longer show up in the appveyor logs,
including those that test very basic and important functionality
(e.g., GitRepo, save, get, and clone).  It looks like these were lost
starting with gh-3975 (2020-01-02), in particular the last commit of
that series (715be1f786).

One change of that commit was to decorate with_testrepos() with
known_failure_appveyor().  (Before the series, the function returned
by with_testrepos() raised a SkipTest if on Windows.)  Based on
replacing the decorator with a plain SkipTest [*], it seems that any
module that uses with_testrepos() isn't executed on appveyor (for
reasons that I don't understand).

Drop the known_failure_appveyor() decorator and instead raise a
SkipTest if on appveyor to restore the execution of these test
modules.

Closes #4411.

[*] https://ci.appveyor.com/project/mih/datalad/builds/35927969