ENH: addurls(..., drop_after=True) to drop content after adding
We have already similar option for add_archive_content and a few places in the
crawler.  It is a very much needed feature while crawling/addurl-ing big
datasets, which might not even fit to local harddrive -- we need to drop
content as soon as it is annexed (well -- later, may be after we extract
metadata).  For now I am just adding it as a bool, but we might later
indeed specify when to drop it (right after added, or after dataset saved
and possibly metadata extracted via some hook or explicit operation etc).

This PR is based on #5022 changes since that is where addurls was RFed
to accomodate parallel execution, and I would prefer to not mess resolving
conflicts later on