ENH: addurl - store original urlfile, not resolved full path
Rationale - to make commit message "portable" happen someone needs to
redo it. Ideally the urlfile should be either "-" or relative within dataset
path.  I decided to not inforce it but merely store the original specification
of the user.

Also decided to just adjust/reuse existing test to not add more time to setup
a dedicated one