BF: make MultiplexSSHConnection.open thread-safe using fasteners.locked
It was reported https://github.com/datalad/datalad/pull/5022#issuecomment-717935610
that parallel install/get over ssh results in flood of

  ControlSocket ... already exists, disabling multiplexing

and abandoned ssh -fN -o ControlMaster=auto ... processes.  This stems from the fact
that MultiplexSSHConnection is nohow thread-safe, so the same instance is used across
multiple threads and races in the check of socket path existing and then acquiring it.

Note: we use threading.Lock here, so this solution is purely for safety across
threads (and more lightweight), and not across multiple processes.  In
principle the construct we have already (not verified to work):

        from .locking import lock_if_check_fails
        with lock_if_check_fails(
            self.ctrl_path.exists,
            self._ctrl_path,
            blocking=False
        ) as (check, lock):
            if check or not (lock and lock.acquired):
                return False
            ... the rest of the .open() call

could be used in addition to establish across process file-based locking in
addition (but would not work for thread safety since uses InterprocessLock).