NF: downloaders: Support downloading Docker registry blobs
An upcoming PR for datalad-container will add a skopeo/OCI-based
adapter for tracking images in a dataset (as an alternative to the
already available docker-save-based adapter).  When the source of the
image is Docker Hub, the layer blobs tracked in the dataset can be
reobtained via a query to the docker.io registry endpoint.

To access Docker Hub registry urls, a token needs to be obtained from
a 401 response for the initial access attempt [1].  The existing
HTTPBearerTokenAuthenticator doesn't fit into this setup because it
expects the token to be known before and supplied by the Credential
class, which isn't provided with the request url.  Add an
authenticator subclass that calls 'HEAD url', grabs the token and
injects it into the header for the 'GET url' call.

[1] https://docs.docker.com/registry/spec/auth/token/