BF: Avoid repo type assumptions to enable proper tracking branch detection
Force-treating a to-be-added submodule as a GitRepo prevented the
detection of a corresponding branch to the active branch, in order to
determine the true tracking branch. Which is necessary to discover
the effective clone URL from a repository (when it is in adjusted
mode).

This is an alternative to 901c8fd4eaf8070fa632f03d9428eceba3c0c97c but
reworked to be applicable to the 0.13 series.

Fixes gh-5107