BF: Fix path semantics in ORA remote
A store's base path was wrongly assumend to be always in POSIX convention,
which prevents operation in non-POSIX platforms.

This change treats the base path as being a PurePosixPath (when derived
from the base URL), but explicitly converts it into a platform Path
prior to any local file system operations.