BF: ora: Trigger SSH cleanup on failure to prevent stall
git-annex's shutdown process for external remotes was tweaked in
7245a9ed5 (Improve shutdown process for external special remotes and
external backends, 2020-11-02), part of the 8.20201103 release.  With
that commit, git-annex waits for stderr rather than closing it at
shutdown of the external remote program.

This is reported to fix a stall on Windows, but it also triggers one
with the ora remote (seen on the github actions with Ubuntu and for me
locally on Debian Buster).  All of the initremote failure checks in
test_initremote_basic() against an SSH target trigger a stall.  This
can be easily triggered outside of the tests by using a non-existing
SSH target:

    $ git annex initremote --debug ria-remote \
      encryption=none type=external externaltype=ora autoenable=true \
      url=ria+ssh://datalad-test/tmp/i-do-not-exist

Even when initremote succeeds before the above commit, ora is leaving
an SSH socket behind.  It tries to register SSHRemoteIO.close(), which
calls SSHManager.close(), as an atexit function, but that doesn't seem
to be in effect here.

Instead of relying on the registered function to get called when
handling errors, call io.close() directly.  This avoids leaving a
socket behind and prevents stalling after git-annex's 7245a9ed5.

Note that close() is only defined for SSHRemoteIO, so this change
comes with additional AttributeError handling.  It might make sense to
add a noop close() method to IOBase, but avoid that for now because
HTTPRemoteIO doesn't actually inherit from IOBase.

Closes #5145.