ENH: datalad.locations.extra-procedures to provide extra location(s) for procedures
This could be useful to point to extra locations for procedures which might
reside in some subdataset, or an alternative to system/user-wide locations.
E.g. it could provide a solution for
https://github.com/datalad/datalad/issues/4802 which avoids need for some
new semantic of --dataset in "create" command, thus keeping it possible to
do

  datalad -c datalad.locations.extra-procedures=$PWD/procedures create  -d . -c mine sub/

as well as

  datalad -c datalad.locations.extra-procedures=$PWD/procedures create  -d ^ -c mine sub/

if current dataset is a subdataset within hierarchy.  Some inconvenience
for use through Python API, for which I filed
https://github.com/datalad/datalad/issues/5155