RF: Avoid run_gitcommand_on_file_list_chunks() (whenever possible)
Calling this wrapper in AnnexRepo._call_annex() is now conditional on
the presence of a `file` argument. This enables (later) support
for connecting `stdin` to an annex process, and generally avoid going
through this constrained wrapper. Whenever desired,
`GitWitlessRunner, run_on_filelist_chunks()` can now be used, which is
a leaner witless-specific reimplementation of the chunking helper.
But this time as a method, and not a standalone function. Its API
is almost identical to that `GitWitlessRunner.run()`.

The decision which interface to use is now made in
AnnexRepo._call_annex() based on the presense of a `file` argument.

This change also avoid gazillions of pointless debug message ala

"Process file list chunk 0"

whenever there only ever is a single chunk.