ENH: addurls: Fall back to addurl when row has empty --key fields
Lift the requirement that all entries have non-empty --key fields.
(This is most relevant for csv/tsv input because, depending on the how
the information is split up across fields, json rows can already fall
back to addurl for specific rows by not including the field keys.)

One downside of this is that --key will have no effect for a format
string with no fields, but there doesn't seem to be a realistic
scenario where that would matter, as it'd mean all files shared the
same key.