BF: drop: Adjust annex's "use --force" message to avoid confusion
Calling `datalad drop` on a file that has too few copies will
confusingly propagate a message from `git annex drop` suggesting to
use --force, an option `datalad drop` does not have.  Replace --force
with --nocheck.

The new message may still be confusing to Python callers of drop()
because it suggests --nocheck rather than check=False, but that's not
a problem unique to this spot and is preferable to reporting the
completely wrong option.

Fixes #5135.