TST/Win: enable removal of long paths on windows via path prefix
shutils rmtree() fails to remove file paths that exceed the 260
character limit. This led to a testfailure in the removal phase of
a temporary annex repo with sha256 backend
(https://github.com/datalad/datalad/pull/5199#issuecomment-737902114).

This change preprends paths under Windows with the magic prefix
'\\?\' to enable the removal:
https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file?redirectedfrom=MSDN#win32-file-namespaces