TST: appveyor: Move plugin tests back to known failures block
On maint, all of the plugin tests are now passing on appveyor
(gh-5202), so 7d01545eb1 moved the plugin tests out of the known
failures block.  Unfortunately tests for master's `addurls --key`
feature bring things back to a failing state (gh-5213).

Until this is resolved, move the plugin tests back to the known
failures block in master's appveyor.yml so that every appveyor run
doesn't need to be checked manually.

[ci skip]