ENH: addurls: Abort if '--key=et:...' is used with old examinekey
The "et:" prefix requires --migrate-to-backend, which was added in
git-annex 8.20201116.  If a caller has an older git-annex and uses the
prefix, we issue a warning and ignore --key.  That's probably not a
good idea because, if the caller specified --key, it's likely they
don't want to download the content.

Abort instead.