BF: addurls --key: Use newer examinekey to support adjusted branches
`git annex fromkey` always tries to create symbolic links [*], so it
won't create the expected file pointers on an adjusted branch and
won't work on file systems that don't support symbolic links.

Check whether a repo is on an adjusted branch and, if it is, use
examinekey's objectpointer field, which isn't available until
8.20201116, to create the pointer file.

On a non-adjusted branch, stick with fromkey because 1) it allows
git-annex to handle more of the operation, including creating
.git/annex/objects/ directories and 2) it's available on all the
git-annex versions we support.

Closes #5213.

[*] It's probably worth asking on git-annex's end whether fromkey
    should write to a pointer file when on an adjusted branch.