TST: github: Loosen string matching of credential exception
test__make_github_repos() is failing on Travis because the the logged
BadCredentialsException ends with

    very bad status "some data"

instead of

    very bad status some data

I see the same failure locally (PyGithub 1.53 and Python 3.7.3).  I'm
hesitant to change it though, because presumably this passed on
@yarikoptic's local machine.

The exact form isn't the crucial part of this test, so just chop off
the trailing part of the query string.

https://travis-ci.com/github/datalad/datalad/jobs/456528059#L2319