BF: Do not raise RuntimeError when there is no JSON output but stderr
This can happen, without being an indication of an error, when I command
like `whereis --json` legitimately produces no results, but at the same
time git-annex runs in debug mode (which can be automatically triggered
via log-level). But what would be an indication of trouble is stdout
content, that wasn't parsed to JSON records.