RF: Use the exact same code path for adding and updating a submodule
For additional improvement TODOs see the diff.

Importantly, this consolidation makes it obsolete to have the result
record be consistent with the one that is yielded by `add()` (which was
used on submodule update). An unfortunate setup, that was criticized
already a long time ago
https://github.com/datalad/datalad/pull/3398#discussion_r282928154

Minus the issue of not breaking the promise of result invariance
over time, we are no free to use sensible result records.