OPT: Report subdataset 'clean' earlier when eval state allows for it
When we have no change in shas, and no worktree modification reported
by `ls-file -m` in the superdataset, a modification can only come from
uncommitted changes. Therefore we can report 'clean' for
`eval_submodule_state==commit`

This optimization causes even less querying of subdatasets, so I had to
adjust the associated test in two ways. The previous method of
"breaking" a subdataset is not effective for `ls-file/ls-tree` which are
the only commands that run now. I switched to an unreadable .git/config
instead. Its effect is now test, because it won't be effective on
windows, and the test will be skipped in that case.