TST: ria: Work around new git-annex check
test_ria_git_remote.test_bare_git_version_{1,2} create a store with a
dataset and check git-annex operation with both the git remote and ora
special remote.  As of git-annex's 75acf5f44 (improve some edge cases
around partial initialization, 2020-12-14) [1], the operations with
the git remote fail because git-annex has a safety check that aborts
when it encounters a repo that has an annex/objects/ directory but no
annex.version or annex.uuid configuration.

The annex/objects/ directory is created by create_ds_in_store().  This
works fine in the context of the create-sibling-ria command: the git
remote has annex-ignore set to true as the intention is for git-annex
to only operate with the store via the ora special remote.  The above
tests, on the other hand, explicitly check git-annex operation on the
git remote.

The git-annex safety check, then, appears to be colliding with a
test-specific detail and shouldn't be an issue with real usage via the
ora special remote.  Given this, update these tests to remove the
annex/objects/ directory after they call create_ds_in_store().

[1] https://git-annex.branchable.com/todo/avoid_autoinit_when_git-annex_branch_already_exists/

Fixes #5253.