BF: annex-ignore RIA sources
The bare repos in a RIA store are intended to not have a local annex
they know of, although there is an annex/ dir for access via ORA special
remote as well as for enabling workflows using ephemeral clones
symlinking into that location.

However, if we clone from local FS, annex-init on the clone may also
initialize origin, posing a danger for messing up availability info.
Prevent that by setting annex-ignore before we initialize the clone (we
used to do that afterwards anyway).

(Closes #5186)