ENH: copy-file: Guard against .git in the target directory
copy-file already does some .git safety checks.  One is that the
destination path that comes out of _yield_specs() doesn't end in .git.
But when there is not an explicit destination path, the target
directory is used, bypassing this check.

Abort if any part of the target directory is ".git".  The handling is
a bit different than the destination paths.  A ValueError is raised
instead of yielding impossible results, because this check can be done
up front on a single value.  Also, with one value, it seems worth the
cost to check all the parts rather than just the final name.