TST: copy-file: Fix .git/ destination test
test_copy_file_prevent_dotgit_placement() has a copy_file() call that
is supposed to return an impossible result because the destination
ends with .git/.  This doesn't work as intended, though.  Due to a
typo, it is actually a list of three items, the last a plain .git
string [*].  The impossible results comes from the "some" source file
not being found.

Fix the typo so that the intended .git/ guard is triggered.

[*] This is treated as a target directory, but the target directory
    code path does not have the same ".git" as explicit destinations.
    The next commit will address that.