RF: Perform checkout of a specific commit right after clone
...and before an `annex init`. Previously this was done in the context
of `get` and the installation of submodules.

This RF enables the proper adjusting of freshly installed subdatasets.
In the previous approach, we would miss and reset an adjusted branch to
the recorded state and essentially break the repository setup, until
the first `save` would rectify it.

This is more or less just shifting code from `get` to `clone`. Please
a bit of renaming and slightly adjusted error handling.

Fixes gh-5257