BF: Restore non-verbose progress logs when non-interactive
When JSON progress by git-annex was handled with
`ProcessAnnexProgressIndicators` there was no progress logging
in non-interactive mode, unless log level was set to 1.

With the switch to `AnnexJsonProtocol` this behavior changed
and now non-interactive logging comes out at the default INFO level.

This change standardizes the non-interactive log level to 5, as used
in other places too.