BF: gitrepo: Update .dirty for adjusted submodule branch handling
This series taught _save_add_submodules() to record the corresponding
branch's id in the superdataset when the submodule is on an adjusted
branch.  `datalad status` is aware of this behavior and ignores
submodule ID changes that are due the a submodule having an adjusted
branch checked out.

This new behavior means that the .dirty optimization from 9d29d13bc9
(OPT: gitrepo.dirty: Use 'git status --porcelain', 2019-05-2) returns
the "wrong" answer for an otherwise clean submodule that is on an
adjusted branch.  Several non-test spots in the code base use .dirty,
so it's important that it doesn't conflict with status().

To resolve this discrepancy, confirm a dirty result reported by git
with the `datalad status` machinery.  When the working tree is clean
in git's eyes, this retains the original speedup, but unfortunately
this may still be enough of a hit to make gh-3342 an issue again.