TST: save: Document issue with adjusted submodules and partial commits
If there are existing staged changes, save_() doesn't stage submodules
changes with _save_add_submodules().  This is wrong because all
submodules need to go through the adjusted-branch-aware handling of
_save_add_submodules().

But it's not just a matter of removing the guard (which this commit
does).  The handling is broken because partial commits are done with
`git commit ... -- FILES`.  This form updates the index entries for
FILES from the working tree.  This undoes any special adjusted branch
handling that _save_add_submodules() did.

Using a temporary index, like the old AnnexRepo.commit() used to do
(removed in bfd9aebf32), is likely the only solution to this.  Given
datalad's "no index" philosophy, it's probably not worth it to address
this, but at least add a test case that demonstrates the issue.