BF: save: Don't overwrite type=dataset,state=clean records
When save() is processing subdatasets, it overrides their status to
type=dataset,status=untracked, except for in the situation covered by
effa32c0d6 (BF: save: Fix handling of explicit paths for nested new
subdatasets, 2020-10-16).  This override is in effect for clean
submodules.

Before the recent change in 20b837d2f0 (RF: Use the exact same code
path for adding and updating a submodule), this resulted in the
unnecessary but quiet action of `git annex add` or `git add` being
called with the submodule.  After that commit, recursively saving a
dataset shows an add record for each subdataset, even if it is clean.

Stop save() from adjusting the status of clean subdatasets so that
_save_add_submodules() is not triggered.