DOC: gitrepo: Decrease duplication among call_git* docstrings
Many of the parameters are shared between the call_git-related
methods.  To avoid repeating them in each docstring, use call_git's
docstring as the primary one, and make the other methods point to it
for shared parameters.

I prefer this approach because it's simple and come with an addition
runtime cost, but another option would be to inject the shared
descriptions, like config.py does with _where_reload.