NF: addurls: Support passing list of records
The urlfile argument takes records via a file, but when addurls is
called from the python interface, specifying an iterable of records
can be more convenient.  Add support for that.

Using urlfile as the name for the positional argument doesn't fit for
this particular use case, but leave it as is to avoiding breaking
existing python callers that use it as a keyword argument.

Closes #4906.