TST: update: Extend comment about a known failure
Of the test_update.py tests, test_update_simple() is now the only test
marked with known_failure_windows and the only test that fails for me
locally under tools/eval_under_testloopfs.

Update the known_failure_windows comment to mention there is an issue
with adjusted branches in general.