TST: update: Make a check compatible with adjusted branches
test_newthings_coming_down() deletes the git-annex branch from a
remote and checks that it is pruned on after an update() call by
asserting an exact set of branches expected is present.  This check
fails if git-annex is in an adjusted state due to the auxiliary
branches.

Tighten to the check to assert what is actually of interest: the
git-annex branch is gone.