TST: Only test line wrapping when we have info on the console width
The test checks against a console width value which can be a guess, and
can be wrong, because our method of determination cannot deal with all
possible circumstances. It would report a default 80, but line wrapping
would still use the actual terminal width when `stdin` is disconnected.

This change disables the length check, when the reported console width
is a guess to avoid failure when the guess is wrong.

Fixes gh-5299