DOC: Include information on subdataset clone sources in `get` docs
So far this information is only available in the docstring of an
internal helper, and in the handbook

http://handbook.datalad.org/en/latest/beyond_basics/101-148-clonepriority.html

However, it is an important piece of info that should be easily
accessible, especially when the `get` of a subdataset has just failed.
This specific use case also motivates, in part, the placement of this
information in the `get` documentation.