TST: rerun: Avoid unnecessary use of --branch
test_rerun_ambiguous_revision_file() uses `branch` when it calls
rerun(), but that's not an essential piece of the test.  With an
upcoming commit, rerun() will abort when `branch` is used with an
adjusted branch, so drop the use of `branch` so that this test is
still executed for file systems where git-annex defaults to an
adjusted branch.