BF: rerun: Abort if using --onto or --branch with adjusted branch
rerun() depends on being able to check out revisions and replay old
commits for --onto and to create a new branch for --branch.  On a file
system where git-annex sets annex.crippledfilesystem to true, doing so
is not conceptually straightforward and would require additional
guards and special handling.  Even on file systems where
annex.crippledfilesystem is _not_ set to true, being on an adjusted
branch would require some care (e.g., not allowing the created branch
to land on an adjusted branch and verifying that the range does not
include an adjusted branch commit).

Regardless of whether any of the above limitations are seen as worth
fixing, rerun() should make the current situation clear to the user.