TST: rerun: Use adjusted-branch compatible commit message inspection
Lots of spots in test_rerun look at the commit message of the last
commit, but that's of course not appropriate when on a _synced_
adjusted branch (and save() syncs the branch after each call).

Import and use the last_commit_msg() helper from test_run, which
inspects DEFAULT_BRANCH (i.e. the corresponding branch for these test
setups).