OPT: Speed-up _check_git_annex_version()
It removes a programmatic assesment of the platform in order to
build a targeted message to be displayed whenever git-annex
cannot be found.

Now the message is generic and saves time for a wide range of
datalad commands executed on the cmdline.

How much is saved is apparently rather variable across systems,
ranging from as little as 30ms to more than 1s (all with Python 3.8
and above, where we use the `distro` package for platform assessment).