TST: Add skip_if_adjusted_branch decorator
A repeated pattern in the tests is to create a dataset with a path
generated via with_{tree,tempfile} and then raise a SkipTest if the
dataset is on an adjusted branch.  Aside from being a bit harder to
grep for, a problem with this pattern is that it's wasteful.  A
dataset is created for each test, but the temporary directory's file
system isn't changing out from under us, so we can use the same answer
across tests.

Add a skip_if_adjusted_branch() decorator that caches the result,
paying the cost of an additional create() call for file systems that
do not default to adjusted branches.

Another advantage of this new decorator is that, when we realize a
test marked with skip_if_on_windows() actually failed due to more
general adjusted branch issues, we can simply switch the decorator.