TST: run: Fix silently failing assertion
As with the test_annexjson_protocol() case from gh-5333, this
assertion has no effect because it happens after the exception is
raised.  Move it outside the assert_raises() context, and fix where we
look for the exit code.

Note that aside from the case fixed in gh-5333, this should be the
last case where an assertion is mistakenly made within the
assert_raises() context.