ENH: add_urls: Use _call_annex_records(..., progress=True)
794dac2da6 (TST: Adjust test for new error destination, 2020-11-19)
tweaked a test in customremotes.tests.test_datalad to look at the
exception string rather than stderr for an add_urls() failure.
However, the assertion actually silently fails because it happens
after the exception in the assert_raises() context.  The string is not
in the exception string because at typical log levels _call_annex() is
called with StdOutCapture; so it's still on stderr, but the test
thinks stderr is empty because swallow_outputs() doesn't work with the
WitlessRunner [1].

It appears that the original reason for making the Runner (and now the
WitlessRunner) not capture stderr was to let wget progress reporting
through.  This may still be effective in some cases (it's doesn't seem
to be on my end), but we can instead use `--json --progress ...`,
going through our more standard progress handling.

Do that by switching to _call_annex_records(..., progress=True).
Update the test to move the now-passing assertion outside of the
assert_raises() context and also drop the use of swallow_outputs
(which has no effect) and of swallow_logs (which isn't used).

[1]: https://github.com/datalad/datalad-container/pull/131#issuecomment-706253763