RF: configuration --scope dataset -> --scope branch
Following the spirit of gh-5361, this change renames the 'dataset'
scope to 'branch', because with a future introduction of other
"dataset" scopes (e.g. git-annex based persistent configuration),
such a generic name will be a problem.

'branch' is a compact and more accurate label.

This changes does not alter labeling in the ConfigManager yet, but
focuses on the about-to-be introduced API addition. Further changes
have less urgency.