ENH(TST): test that we run ok in a forked process as well
Got a bit complicated to avoid child keep running the other tests, and we could not
just sys.exit(0) from the child - then parent process would not know to mark it as a failure.
So we really needed to kill the child.

I really hope that OS/Python/my terminology will not trigger interest of any
child abuse services/initiatives, and demand Python to rename "kill" to "letgo"
or something alike to not impose some violent behavior in our code.