MNT: CHANGELOG.md: Revise 0.14.0 entries
Revert a change from 9e111baa2a that was mistakenly tweaked a 0.13.6
entry, adding a new entry for 0.14.0.

The "{main => primary} branch" change is to avoid any confusion with
the default "main" branch now used by some hosting sites.
"Corresponding" (which we use in the code base) would be another
option, but for some reason it seems less clear to me than "primary"
in this context.